Support simtrace logs from s3 in log analysis notebook
Previously the simtrace logs were logged to cloudwatch and the log analysis notebook used these logs for analysing the model. Instead of populting the logs with simtrace all the simtrace are uploaded to S3 bucket.

To support this I made modification to the notebook. Also the log analysis notebook was strongly coupled to the reInvent 2018 track. I have made modification to work with other track.